WECS-157 : Move Pager to a new component and make minor adjustments
